+++ Gardener-main/RuleEngine/RuleEngine.hpp
@@ -2,12 +2,16 @@
 #define RULEENGINE_HPP
 #include <string>
 #include <functional>
+#include <initializer_list>
+#include <vector>
 #include <Print.h>
 #include "../src/TinyExpr/tinyexpr.h"
 #include "ArduinoJson.h"
 #include "property.hpp"
 #include "propertystore.hpp"
 #include "../Debug/Debug.hpp"
+#include "Rule.hpp"
+#include "VectorBaseStore/VectorBaseStore.hpp"
 
 /// @brief Helper function to extract a te_type type value from a templated `Property` instance.
 /// @tparam T The type of the `Property`, usually `bool`, `int32_t`, or `float`.
@@ -53,6 +57,10 @@ te_type get_property_value(const te_expr* context)
 class RuleEngine 
 {
     public:
+        /// @brief 
+        /// @param debugger The debug interface to print messages to.
+        RuleEngine(std::initializer_list<Rule*> rules_list, Debug& debugger);
+
         ~RuleEngine();
 
         /// @brief Add a set of `Property`s to the expression parser.
@@ -65,7 +73,7 @@ class RuleEngine
         
         /// @brief End of the recursive `set_variables` function.
         void set_variables();
-        
+
         /// @brief Start the engine
         void begin();
 
@@ -77,25 +85,30 @@ class RuleEngine
         void process_attributes(JsonObject obj);
 
         /// @brief Runs the engine. Should be implemented in derived classes.
-        virtual void loop() = 0;
+        virtual void loop();
         
         /// @brief Print information about the rule engine to the `debug`er passed to the constructor.
-        virtual void print() = 0;
+        virtual void print();
 
-        virtual BaseStore& get_rules() = 0;
+        virtual BaseStore& get_rules();
+
+        /// @brief Process a command.
+        /// @param cmd 
+        virtual void process_command(String& cmd);
         
     protected:
-        /// @brief 
-        /// @param debugger The debug interface to print messages to.
-        RuleEngine(Debug* debugger);
-
         /// @brief Add or update a rule with the parameters contained in `pair`.
         /// @param pair The parameters needed by the rule.
         /// @return true, if the rule was constructed successfully.
-        virtual bool process_rule(JsonPair pair) = 0;
+        virtual bool process_rule(JsonPair pair);
         
         /// @brief All the variables have been set, so the rules can now be compiled.
-        virtual void compile_rules() = 0;
+        virtual void compile_rules();
+
+        /// @brief Find rule by name
+        /// @param name The name
+        /// @return The rule if found, otherwise a `nullptr`.
+        virtual Rule* find_rule(const char* name);
 
         /// @brief A base expression parser that is provided with the variables, and is copied to all 
         /// rules to be compiled with an expression.
@@ -106,10 +119,17 @@ class RuleEngine
         
         /// @brief A pointer to a `Debug` object that prints debug messages to the relevant outputs.
         Debug* debug;
+
+        /// @brief The set of rules to evaluate.
+        std::vector<Rule*> rules;
+
+        VectorBaseStore rules_basestore;
 };
 
-RuleEngine::RuleEngine(Debug* debugger):
-    debug(debugger)
+RuleEngine::RuleEngine(std::initializer_list<Rule*> rules_list, Debug& debugger):
+    rules(rules_list),
+    rules_basestore(rules),
+    debug(&debugger)
 {}
 
 RuleEngine::~RuleEngine()
@@ -159,5 +179,144 @@ void RuleEngine::process_attributes(JsonObject obj)
     }
 }
 
+void RuleEngine::loop()
+{
+    for(Rule* rule : rules)
+    {
+        if(rule->enabled)
+        {
+            // Its time to evaluate one
+            if(millis() - rule->last_evaluation >= (rule->eval_interval * 1000ul))
+            {
+                rule->last_evaluation = millis();
+                debug->printv("[RuleEngine] evaluating ", rule->get_name());
+                rule->update();
+            }
+        }
+    }
+}
+
+bool RuleEngine::process_rule(JsonPair pair)
+{
+    // JsonPair pair needs to have the following structure:
+    // "rule_id": {
+    //     "expression": "gerrit + henk == chaos",
+    //     "eval_interval": 600,
+    //     "enabled": true
+    //     "address": 25
+    // }
+    
+    const char* rule_name = pair.key().c_str();
+
+    if(!pair.value().is<JsonObject>())
+    {
+        // Value is not a JsonObject.
+        debug->printv("[RuleEngine] ERROR: Value is not a JsonObject.");
+        return false;
+    }
+    
+    JsonObject params = pair.value();
+    for(Rule* rule : rules)
+    {
+        // Found rule
+        if(strcmp(rule->name, rule_name) == 0)
+        {
+            debug->printv("[RuleEngine] Found existing rule with name ", rule_name);
+            rule->set_from_json(params);
+            rule->set_parser(base_parser);
+            rule->compile();
+
+            if(!rule->compiled)
+            {
+                // Expression compilation error.
+                debug->printv("[RuleEngine] ERROR: Expression compilation failed");
+                debug->printv(rule->parser.get_last_error_position());
+                debug->printv(rule->parser.get_last_error_message().c_str());
+                return false;
+            }
+            else
+            {
+                debug->printv("[RuleEngine] Rule updated");
+                return true;
+            }
+        }
+    }
+    debug->printv("[RuleEngine] Rule not found");
+    return false;
+}
+
+void RuleEngine::compile_rules()
+{
+    for(Rule* rule : rules)
+    {
+        rule->set_parser(base_parser);
+        rule->compile();
+    }
+}
+
+Rule *RuleEngine::find_rule(const char *name)
+{
+    for(Rule* rule : rules)
+    {
+        // Rule already exists.
+        if(strcmp(rule->name, name) == 0)
+        {
+            return rule;
+        }
+    }
+    return nullptr;
+}
+
+void RuleEngine::print()
+{
+    debug->printv("[RuleEngine] with ", rules.size(), " rules.");
+
+    for(Rule* rule : rules)
+    {
+        rule->print_to(*debug);
+    }
+}
+
+BaseStore& RuleEngine::get_rules()
+{
+    return rules_basestore;
+}
+
+void RuleEngine::process_command(String &cmd)
+{
+    int var_sep = cmd.indexOf(':');
+    if(var_sep > 0)
+    {
+        String name = cmd.substring(0, var_sep);
+        String val  = cmd.substring(var_sep + 1);
+        if(name == "eval")
+        {
+            // Force evaluation of rule identified by val
+            Rule* rule = find_rule(val.c_str());
+            if(rule != nullptr)
+            {
+                debug->printv("[RuleEngine] evaluating ", rule->get_name());
+                rule->update();
+            }
+            return;
+        }
+        // Modify a rule
+        else if(name == "set_rule")
+        {
+            JsonDocument doc;
+            DeserializationError err = deserializeJson(doc, val);
+            if(err)
+            {
+                debug->printv("[RuleEngine] ERROR parsing JSON");
+                debug->printv(err.c_str());
+            }
+            else
+            {
+                debug->printv("[RuleEngine] processing rule.");
+                process_attributes(doc.as<JsonObject>());
+            }
+        }
+    }
+}
 
 #endif